feat(journey-maps): code review: merge classes
Refs: ROKAS-1204
