+++ src/journey-maps/angular/services/map/util/style-version-lookup.ts
