+++ src/journey-maps/angular/services/map/util/walk-floor-updater.ts
